Now using new scripts to build & start electron/local web
