+++ package.json
@@ -5,8 +5,8 @@
   "main": "electron/index.js",
   "scripts": {
     "start": "electron .",
-    "start-electron-local": "npm run build-styles && node ./bin/copyThemes.js && npm start",
-    "start-web": "node ./bin/deleteThemes.js && ui5 serve",
+    "start-electron-local": "npm run build-styles && node ./bin/prepareForElectron.js && npm start",
+    "start-web": "node ./bin/prepareForWeb.js && ui5 serve",
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "npm run ui5-build",
     "ui5-build": "ui5 build -a self-contained --dest=dist-ui5 && npm run build-styles && node ./bin/copyThemesDist.js",
@@ -41,7 +41,8 @@
     "chokidar": "^3.4.0",
     "electron": "^8.5.2",
     "electron-builder": "^22.9.1",
-    "ui5-middleware-index": "^0.2.2"
+    "ui5-middleware-index": "^0.2.2",
+    "chalk": "^4.1.0"
   },
   "dependencies": {
     "electron-store": "^5.1.1",
